backend/requirements.txt
Flask>=2.0.0
werkzeug >= 2.0.0
requests
psycopg2-binary
python-dotenv
redis